vlan deployment script
This script will prompt the user in the terminal about which vlan they want to create. Then it will connect to the device and verify if the vlan exists or not. If it doesn't exist than the vlan will be created. If not then it will send a prompt to the user advising the vlan already exists.